# webui-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaddleOCR Local WebUI",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/model-runtime/switch": {
      "post": {
        "summary": "Switch Model Runtime",
        "operationId": "switch_model_runtime_api_model_runtime_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/model-runtime/deploy": {
      "post": {
        "summary": "Deploy Model Runtime",
        "operationId": "deploy_model_runtime_api_model_runtime_deploy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/unlimited-ocr/backend": {
      "get": {
        "summary": "Get Unlimited Ocr Backend",
        "operationId": "get_unlimited_ocr_backend_api_unlimited_ocr_backend_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Switch Unlimited Ocr Backend",
        "operationId": "switch_unlimited_ocr_backend_api_unlimited_ocr_backend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlimitedOcrBackendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelDeployRequest": {
        "properties": {
          "modelId": {
            "type": "string",
            "title": "Modelid"
          },
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          }
        },
        "type": "object",
        "required": [
          "modelId"
        ],
        "title": "ModelDeployRequest"
      },
      "ModelSwitchRequest": {
        "properties": {
          "modelId": {
            "type": "string",
            "title": "Modelid"
          }
        },
        "type": "object",
        "required": [
          "modelId"
        ],
        "title": "ModelSwitchRequest"
      },
      "UnlimitedOcrBackendRequest": {
        "properties": {
          "backend": {
            "type": "string",
            "title": "Backend"
          }
        },
        "type": "object",
        "required": [
          "backend"
        ],
        "title": "UnlimitedOcrBackendRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}